+++ src/projects/index.js
@@ -1,5 +1,6 @@
-export { ComingSoon } from './ComingSoon';
-export { NotBopIt } from './NotBopIt';
-export { PythonRedex } from './PythonRedex';
-export { TextClassifier } from './TextClassifier';
-export { Website } from './Website';
+export { ComingSoon } from "./ComingSoon";
+export { NotBopIt } from "./NotBopIt";
+export { PythonRedex } from "./PythonRedex";
+export { TextClassifier } from "./TextClassifier";
+export { Website } from "./Website";
+export { IntermittentComputing } from "./IntermittentComputing";