+++ CHANGES.md
@@ -59,3 +59,7 @@
 
 - Fixed background gradient
 - Updated colors
+
+**11/1/2024**
+
+- Added intermittent computing project